more scripts

+++ certificates/scripts/proofs/slurmWrapperEvaluate.sh
@@ -6,6 +6,8 @@ output_file=$3
 
 if ! [[ $# -eq 3 ]]; then
     echo "Usage: $0 <input dir> <output dir> <output file base name>"
+    echo "proof will be stored in output dir. Log data will be stored in output_file_base_name.json and
+    output_file_base_name.csv in the current directory"
     exit 1
 fi
 
@@ -26,6 +28,7 @@ output_file_json=$output_file".json"
 echo "[" > $output_file_json
 output_file_csv=$output_file".csv"
 rm -f $output_file_csv
+
 
 while read -r output_log; do
   while IFS= read -r line; do
@@ -49,10 +52,9 @@ while read -r output_log; do
           ;; 
         json)
 	  json=$value
-	  echo $json"," >> $output_file_json
           ;;
         *)
-          echo "Unknown key: $key → $value"
+          #echo "Unknown key: $key → $value"
           ;;
       esac
       ;;
@@ -63,8 +65,14 @@ while read -r output_log; do
   then
     mkdir -p $output_dir/$rel_path
     directory=$(dirname "$output_log")
-    cp $directory/$proof_name $output_dir/$rel_path/
-    echo "$problem_path,$output_dir/$rel_path/$proof_name" >> $output_file_csv
+    if [[ -e "$directory/$proof_name" ]]
+    then
+      cp $directory/$proof_name $output_dir/$rel_path/
+      echo "$problem_path,$output_dir/$rel_path/$proof_name" >> $output_file_csv
+      echo $json"," >> $output_file_json
+    else
+      echo "Could not copy $directory/$proof_name"
+    fi
   fi
 done <<< $(find "$input_dir" -type f -name "output.log")
 
@@ -72,3 +80,5 @@ done <<< $(find "$input_dir" -type f -name "output.log")
 sed -i '$s/,$//' "$output_file_json"
 echo "]" >> $output_file_json
 
+mkdir -p $output_dir/old_runs
+cp $output_file_json $output_dir/old_runs/"log_$(date +%Y-%m-%d_%H-%M-%S).json"
